Refactor translation structure for waste management plugin
- Updated import paths for translation files to use new shared modules.
- Created separate shared modules for master data, scheduling, and sections translations.
- Removed the old shared translations file to streamline the codebase.
- Adjusted tests to reflect the new structure and imports.
