+++ apps/sva-studio-react/src/routes/admin/instances/-instances-shared.tsx
@@ -7,13 +7,8 @@ import { Badge } from '../../../components/ui/badge';
 import { t } from '../../../i18n';
 import type { IamHttpError } from '../../../lib/iam-api';
 import type {
-  InstanceConfigurationOverallStatus,
-  WorkflowStepState,
-} from './-instances-shared-types';
-import type {
   EvidenceSource,
   HistoryWorkspaceModel,
-  OperationStepStatus,
   OperationsDetailAction,
   OperationsPrimaryAction,
   OperationsStepKey,
@@ -21,7 +16,7 @@ import type {
   RealmOperationsModel,
 } from './-instance-detail-operations-types';
 import { getInstanceErrorMessage } from './-instance-error-message-shared';
-import { translateConfigurationStatus, findPreflightCheck } from './-instance-detail-shared';
+import { findPreflightCheck } from './-instance-detail-shared';
 
 export type {
   CockpitAnomalyItem,
@@ -34,7 +29,6 @@ export type {
   InstanceConfigurationOverallStatus,
   InstanceDetailCockpitModel,
   InstanceFieldHelpKey,
-  PostCreateGuidanceInput,
   PrimaryDetailAction,
   SelectedInstance,
   SetupWorkflowStep,
@@ -54,16 +48,7 @@ export type {
   RealmOperationsModel,
 } from './-instance-detail-operations-types';
 
-export type DetailWorkspaceTab = 'overview' | 'configuration' | 'history';
 export const getErrorMessage = getInstanceErrorMessage;
-
-const OPERATION_STATUS_BADGE_LABELS: Record<OperationStepStatus, string> = {
-  offen: 'admin.instances.operations.status.offen',
-  bereit: 'admin.instances.operations.status.bereit',
-  läuft: 'admin.instances.operations.status.laeuft',
-  erfolgreich: 'admin.instances.operations.status.erfolgreich',
-  fehlgeschlagen: 'admin.instances.operations.status.fehlgeschlagen',
-};
 
 const NEW_REALM_STEP_TITLES: Record<
   Exclude<OperationsStepKey, 'live_status' | 'drift_analysis' | 'contract_repair' | 'reconcile' | 'result_validation'>,
@@ -732,31 +717,6 @@ export const buildHistoryWorkspaceModel = (
     historicalRuns,
     hasHistoricalMismatchHint,
   };
-};
-
-export const KeycloakStatusBadge = ({ ready }: { ready: boolean }) => (
-  <Badge variant={ready ? 'secondary' : 'outline'}>
-    {ready ? t('admin.instances.keycloakStatus.ok') : t('admin.instances.keycloakStatus.missing')}
-  </Badge>
-);
-
-export const ConfigurationStatusBadge = ({ status }: { status: InstanceConfigurationOverallStatus }) => {
-  const variant = status === 'complete' ? 'secondary' : 'outline';
-  return <Badge variant={variant}>{translateConfigurationStatus(status)}</Badge>;
-};
-
-export const WorkflowStatusBadge = ({ status }: { status: WorkflowStepState }) => {
-  const labelMap: Record<WorkflowStepState, string> = {
-    done: t('admin.instances.workflow.badges.done'),
-    current: t('admin.instances.workflow.badges.current'),
-    blocked: t('admin.instances.workflow.badges.blocked'),
-    pending: t('admin.instances.workflow.badges.pending'),
-  };
-  return <Badge variant={status === 'done' ? 'secondary' : 'outline'}>{labelMap[status]}</Badge>;
-};
-
-export const OperationsStepStatusBadge = ({ status }: { status: OperationStepStatus }) => {
-  return <Badge variant={status === 'erfolgreich' ? 'secondary' : 'outline'}>{t(OPERATION_STATUS_BADGE_LABELS[status])}</Badge>;
 };
 
 export const ProvisioningStepBadge = ({
