+++ knip.json
@@ -29,6 +29,13 @@
     "apps/sva-studio-react/vitest.ui.config.ts",
     "apps/project-report/vitest.config.ts",
     "packages/*/vitest.config.ts",
+    "packages/auth-runtime/src/iam-instance-registry/core-runtime.ts",
+    "packages/auth-runtime/src/iam-instance-registry/index.ts",
+    "packages/auth-runtime/src/iam-instance-registry/service-detail.ts",
+    "packages/auth-runtime/src/iam-instance-registry/service-keycloak-execution-shared.ts",
+    "packages/auth-runtime/src/iam-instance-registry/service-keycloak-execution.ts",
+    "packages/auth-runtime/src/iam-instance-registry/service.ts",
+    "packages/auth-runtime/src/iam-instance-registry/worker.ts",
     "packages/data/scripts/**",
     "packages/media/src/*.type-test.ts",
     "packages/monitoring-client/scripts/**",
