+++ packages/auth-runtime/src/iam-account-management/schemas.ts
@@ -113,24 +113,3 @@ export const updateRoleSchema = z
     (value) => Object.keys(value).length > 0 && (Object.keys(value).some((key) => key !== 'retrySync') || value.retrySync),
     'Mindestens ein Feld muss gesetzt werden.'
   );
-
-export const createGroupSchema = z.object({
-  groupKey: z
-    .string()
-    .trim()
-    .min(3)
-    .max(64)
-    .regex(/^[a-z0-9_]+$/),
-  displayName: z.string().trim().min(1).max(120),
-  description: z.string().trim().max(500).optional(),
-  roleIds: uniqueUuidArray(50).default([]),
-});
-
-export const updateGroupSchema = z
-  .object({
-    displayName: z.string().trim().min(1).max(120).optional(),
-    description: z.string().trim().max(500).optional(),
-    roleIds: uniqueUuidArray(50).optional(),
-    isActive: z.boolean().optional(),
-  })
-  .refine(hasDefinedEntries, 'Mindestens ein Feld muss gesetzt werden.');
