--- packages/auth-runtime/src/iam-account-management/user-detail-query.mapping.ts
@@ -1,1 +1,0 @@
-export { mapUserDetailRow } from '@sva/iam-admin';
