--- packages/auth-runtime/src/iam-account-management/user-detail-query.permission-rows.ts
@@ -1,4 +1,0 @@
-export {
-  buildDirectPermissionRowsSql,
-  buildPermissionRowsSql,
-} from '@sva/iam-admin';
