--- packages/auth-runtime/src/iam-account-management/user-detail-query.permission-trace.ts
@@ -1,1 +1,0 @@
-export { buildPermissionTraceRowsSql } from '@sva/iam-admin';
