--- packages/auth-runtime/src/iam-account-management/user-detail-query.sql.ts
@@ -1,4 +1,0 @@
-export {
-  readUserDetailSchemaSupport,
-  selectUserDetailQuery,
-} from '@sva/iam-admin';
