--- packages/auth-runtime/src/iam-account-management/user-detail-query.types.ts
@@ -1,9 +1,0 @@
-export type {
-  UserDetailDirectPermissionRow,
-  UserDetailGroupRow,
-  UserDetailPermissionTraceRow,
-  UserDetailRoleRow,
-  UserDetailRow,
-  UserDetailSchemaSupport,
-  UserDetailSchemaSupportRow,
-} from '@sva/iam-admin';
