--- packages/auth-runtime/src/iam-account-management/user-list-query.ts
@@ -1,1 +1,0 @@
-export { resolveUsersWithPagination } from '@sva/iam-admin';
