+++ packages/plugin-waste-management/src/plugin.translations.de.masterData.ts
@@ -1,5 +1,8 @@
 import { wasteManagementPluginTranslationsDEMasterDataLocationsWorkspace } from './plugin.translations.de.masterData.locations-workspace.js';
-import { createMasterDataTabs, createWasteManagementMasterDataTranslations } from './plugin.translations.shared.js';
+import {
+  createMasterDataTabs,
+  createWasteManagementMasterDataTranslations,
+} from './plugin.translations.shared.master-data.js';
 
 export const wasteManagementPluginTranslationsDEMasterData = createWasteManagementMasterDataTranslations({
   meta: {
