+++ packages/plugin-waste-management/src/plugin.translations.de.overview.ts
@@ -1,4 +1,4 @@
-import { createWasteManagementOverviewTranslations } from './plugin.translations.shared.js';
+import { createWasteManagementOverviewTranslations } from './plugin.translations.shared.sections.js';
 
 export const wasteManagementPluginTranslationsDEOverview = createWasteManagementOverviewTranslations({
   messages: {
