+++ packages/plugin-waste-management/src/plugin.translations.de.settings.ts
@@ -1,4 +1,4 @@
-import { createWasteManagementSettingsTranslations } from './plugin.translations.shared.js';
+import { createWasteManagementSettingsTranslations } from './plugin.translations.shared.sections.js';
 
 export const wasteManagementPluginTranslationsDESettings = createWasteManagementSettingsTranslations({
   groupTitle: 'Waste-Datenquelle',
