+++ packages/plugin-waste-management/src/plugin.translations.de.tools.ts
@@ -1,4 +1,4 @@
-import { createWasteManagementToolsTranslations } from './plugin.translations.shared.js';
+import { createWasteManagementToolsTranslations } from './plugin.translations.shared.sections.js';
 
 export const wasteManagementPluginTranslationsDETools = createWasteManagementToolsTranslations({
   imports: {
