+++ packages/plugin-waste-management/src/plugin.translations.en.scheduling.ts
@@ -1,4 +1,4 @@
-import { createWasteManagementSchedulingTranslations } from './plugin.translations.shared.js';
+import { createWasteManagementSchedulingTranslations } from './plugin.translations.shared.scheduling.js';
 
 export const wasteManagementPluginTranslationsENScheduling = createWasteManagementSchedulingTranslations({
   global: {
